Template inheritance
This is template inheritance. Here we can see that there is one base.html which is the base for all the html, this means we will have to type the entire html only once. The "jinjutsu" code in the base html are variables and can be changed depending on which page is calling the base, in this case it is the index.html.